+++ dtale/cli/loaders/csv_loader.py
@@ -1,5 +1,6 @@
 import pandas as pd
 
+from dtale.app import show
 from dtale.cli.clickutils import get_loader_options
 
 '''
@@ -11,6 +12,16 @@ LOADER_PROPS = [
     dict(name='path', help='path to CSV file'),
     dict(name='parse_dates', help='comma-separated string of column names which should be parsed as dates')
 ]
+
+
+# IMPORTANT!!! This function is required if you would like to be able to use this loader from the back-end.
+def show_loader(**kwargs):
+    return show(data_loader=lambda: loader_func(**kwargs), **kwargs)
+
+
+def loader_func(**kwargs):
+    path = kwargs.pop('path')
+    return pd.read_csv(path, **{k: v for k, v in kwargs.items() if k in LOADER_PROPS})
 
 
 # IMPORTANT!!! This function is required for building any customized CLI loader.
@@ -28,7 +39,7 @@ def find_loader(kwargs):
             csv_arg_parsers = {  # TODO: add additional arg parsers
                 'parse_dates': lambda v: v.split(',') if v else None
             }
-            kwargs = {k: csv_arg_parsers.get(k, lambda v: v)(v) for k, v in csv_opts.items() if k != 'path'}
-            return pd.read_csv(csv_opts['path'], **kwargs)
+            kwargs = {k: csv_arg_parsers.get(k, lambda v: v)(v) for k, v in csv_opts.items()}
+            return loader_func(**kwargs)
         return _csv_loader
     return None