+++ dtale/cli/loaders/json_loader.py
@@ -1,5 +1,7 @@
 import pandas as pd
+import requests
 
+from dtale.app import show
 from dtale.cli.clickutils import get_loader_options
 
 '''
@@ -11,6 +13,26 @@ LOADER_PROPS = [
     dict(name='path', help='path to JSON file or URL to JSON endpoint'),
     dict(name='convert_dates', help='comma-separated string of column names which should be parsed as dates')
 ]
+
+
+# IMPORTANT!!! This function is required if you would like to be able to use this loader from the back-end.
+def show_loader(**kwargs):
+    return show(data_loader=lambda: loader_func(**kwargs), **kwargs)
+
+
+def loader_func(**kwargs):
+    path = kwargs.pop('path')
+    normalize = kwargs.pop('normalize', False)
+    if path.startswith('http://') or path.startswith('https://'):  # add support for URLs
+        proxy = kwargs.pop('proxy', None)
+        req_kwargs = {}
+        if proxy is not None:
+            req_kwargs['proxies'] = dict(http=proxy, https=proxy)
+        resp = requests.get(path, **req_kwargs)
+        path = resp.json() if normalize else resp.text
+    if normalize:
+        return pd.io.json.json_normalize(path, **kwargs)
+    return pd.read_json(path, **{k: v for k, v in kwargs.items() if k in LOADER_PROPS})
 
 
 # IMPORTANT!!! This function is required for building any customized CLI loader.
@@ -28,7 +50,7 @@ def find_loader(kwargs):
             json_arg_parsers = {  # TODO: add additional arg parsers
                 'parse_dates': lambda v: v.split(',') if v else None
             }
-            kwargs = {k: json_arg_parsers.get(k, lambda v: v)(v) for k, v in json_opts.items() if k != 'path'}
-            return pd.read_json(json_opts['path'], **kwargs)
+            kwargs = {k: json_arg_parsers.get(k, lambda v: v)(v) for k, v in json_opts.items()}
+            return loader_func(**kwargs)
         return _json_loader
     return None