+++ dtale/global_state.py
@@ -4,6 +4,7 @@ DTYPES = {}
 SETTINGS = {}
 METADATA = {}
 CONTEXT_VARIABLES = {}
+HISTORY = {}
 
 
 def get_data(data_id=None):
@@ -46,6 +47,14 @@ def get_context_variables(data_id=None):
     return CONTEXT_VARIABLES.get(data_id)
 
 
+def get_history(data_id=None):
+    global HISTORY
+
+    if data_id is None:
+        return HISTORY
+    return HISTORY.get(data_id)
+
+
 def set_data(data_id, val):
     global DATA
 
@@ -76,17 +85,28 @@ def set_context_variables(data_id, val):
     CONTEXT_VARIABLES[data_id] = val
 
 
-def cleanup():
+def set_history(data_id, val):
+    global HISTORY
+
+    HISTORY[data_id] = val
+
+
+def cleanup(data_id=None):
     """
     Helper function for cleanup up state related to a D-Tale process with a specific port
 
     :param port: integer string for a D-Tale process's port
     :type port: str
     """
-    global DATA, DTYPES, SETTINGS, METADATA, CONTEXT_VARIABLES
+    global DATA, DTYPES, SETTINGS, METADATA, CONTEXT_VARIABLES, HISTORY
 
-    DATA = {}
-    SETTINGS = {}
-    DTYPES = {}
-    METADATA = {}
-    CONTEXT_VARIABLES = {}
+    if data_id is None:
+        DATA = {}
+        SETTINGS = {}
+        DTYPES = {}
+        METADATA = {}
+        CONTEXT_VARIABLES = {}
+        HISTORY = {}
+    else:
+        for data in [DATA, DTYPES, SETTINGS, METADATA, CONTEXT_VARIABLES, HISTORY]:
+            data.pop(data_id, None)  # use dict.pop with a default so that KeyError won't occur